Remove Insulin Fraction Warning from Dynamic CR Verbose Hint
Removes "It is not recommended to use this feature with an insulin fraction >2" Warning.

There is no setting labeled "Insulin Fraction" and no use of "Insulin Fraction" in Trio 1.0 dosing calculations. After reviewing all possibilities of what this could reference and determining that all possible settings are limited to a maximum of 2 or less, this commit serves to remove this warning to avoid unnecessary user confusion.