Swift port of the autosens algorithm
In this current PR we port the Javascript autosens algorithm to
swift. The JS implementation is in a single monolithic function, so
for testing we captured JSON logs from a real device and use these in
swift to test. We added a few unit tests to cover the major cases not
included in our real logs.