Update CoB bucketing algorithm to match JS
This commit includes an updated CoB algorithm that implements the JS
logic faithfully. The big changes include using the JS bucketing
algorithm and mutating inputs for the CoB algorithm.

It also includes updates to the unit tests to accommodate the new code
and updates after fixing the JS unit tests.