First cut at carbsReq
This commit required a bit of rearchitecting to pass the right values
around to the right place. In particular, I had to make a number of
changes to the forecasting logic for the min/max glucose calculations
to make sure that it matches JS for downstream computations.

Also, there are a few precision / rounding updates to better match JS