Early exit tests
This commit adds test cases to check all of the ways the determine
basal function can exit early before hitting the core dosing
logic. There are corresponding Javascript tests that show that these
are the right conditions to check. It also includes code to wrap up
these early exits into error objects.

We removed the `enableSMB` variable in the Determination that we used
for debugging and updated the determineBasal JS to remove it from
there as well.