Ensure consistency for determineBasal `result` strings
This commit includes a number of small fixes to make the Swift reason
string construction consistent with Javascript.

We did have to update the testing JS to:
- Add precision to maxSafeBasalRate
- Add rounding to a value in the result string
- Remove contributions from `insulinForManualBolus`, which Trio doesn't support

And there is one tiny bug fix in the Swift implementation around our
use of `profile.sens` where we should always try to get that first and
then fall back to the lookup based on time. The issue is that the time
for profile vs determine basal might cross a sensitivity boundary.

Finally, this commit adds a few properties to our fuzzy matching logic
that Trio uses and has some small fixes to ensure consistency.